refactor(BeanScanner): 일반화 강화
